cloud_functions/load_to_warehouse/requirements.txt
functions-framework==3.*
Flask==2.0.2
google-cloud-storage==2.9.0
